materials/05-shiny-app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via
    #   pins
    #   shiny
asgiref==3.8.1
    # via shiny
asttokens==2.4.1
    # via stack-data
atpublic==4.1.0
    # via ibis-framework
bidict==0.23.1
    # via ibis-framework
branca==0.7.2
    # via ipyleaflet
build==1.2.1
    # via pip-tools
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pip-tools
    #   rsconnect-python
    #   shiny
    #   typer
    #   uvicorn
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
debugpy==1.8.2
    # via ipykernel
decorator==5.1.1
    # via ipython
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
executing==2.0.1
    # via stack-data
fastapi==0.111.0
    # via vetiver
fastapi-cli==0.0.4
    # via fastapi
fsspec==2024.6.0
    # via pins
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
htmltools==0.5.2
    # via shiny
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   vetiver
humanize==4.9.0
    # via pins
ibis-framework==9.1.0
    # via -r requirements.in
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==7.1.0
    # via pins
importlib-resources==6.4.0
    # via pins
ipykernel==6.29.5
    # via -r requirements.in
ipyleaflet==0.19.1
    # via -r requirements.in
ipython==8.18.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.1.3
    # via
    #   ipyleaflet
    #   shinywidgets
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   branca
    #   fastapi
    #   pins
joblib==1.4.2
    # via
    #   pins
    #   scikit-learn
    #   vetiver
jupyter-client==8.6.2
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   shinywidgets
jupyter-leaflet==0.19.1
    # via ipyleaflet
jupyterlab-widgets==3.0.11
    # via ipywidgets
linkify-it-py==2.0.3
    # via shiny
loguru==0.7.2
    # via -r requirements.in
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   shiny
markupsafe==2.1.5
    # via jinja2
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdit-py-plugins==0.4.1
    # via shiny
mdurl==0.1.2
    # via markdown-it-py
nest-asyncio==1.6.0
    # via
    #   ipykernel
    #   vetiver
numpy==1.26.4
    # via
    #   ibis-framework
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   vetiver
orjson==3.10.6
    # via fastapi
packaging==24.0
    # via
    #   build
    #   htmltools
    #   ipykernel
    #   plotly
    #   shiny
pandas==2.2.2
    # via
    #   ibis-framework
    #   pins
    #   vetiver
parso==0.8.4
    # via jedi
parsy==2.1
    # via ibis-framework
pexpect==4.9.0
    # via ipython
pins==0.8.6
    # via
    #   -r requirements.in
    #   vetiver
pip==24.1.1
    # via
    #   pip-tools
    #   rsconnect-python
pip-tools==7.4.1
    # via vetiver
platformdirs==4.2.2
    # via jupyter-core
plotly==5.22.0
    # via
    #   -r requirements.in
    #   vetiver
polars==0.20.31
    # via -r requirements.in
prompt-toolkit==3.0.36
    # via
    #   ipython
    #   questionary
psutil==6.0.0
    # via ipykernel
psycopg2==2.9.9
    # via ibis-framework
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyarrow==16.1.0
    # via
    #   -r requirements.in
    #   ibis-framework
pyarrow-hotfix==0.6
    # via ibis-framework
pydantic==2.8.2
    # via
    #   fastapi
    #   vetiver
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via
    #   ipython
    #   rich
pyjwt==2.8.0
    # via rsconnect-python
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   ibis-framework
    #   jupyter-client
    #   pandas
    #   shinywidgets
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
    #   vetiver
python-multipart==0.0.9
    # via
    #   fastapi
    #   shiny
pytz==2024.1
    # via
    #   ibis-framework
    #   pandas
pyyaml==6.0.1
    # via
    #   pins
    #   uvicorn
pyzmq==26.0.3
    # via
    #   ipykernel
    #   jupyter-client
questionary==2.0.1
    # via shiny
requests==2.32.3
    # via
    #   pins
    #   vetiver
rich==13.7.1
    # via
    #   ibis-framework
    #   typer
rsconnect-python==1.24.0
    # via vetiver
scikit-learn==1.5.1
    # via vetiver
scipy==1.14.0
    # via scikit-learn
semver==2.13.0
    # via rsconnect-python
setuptools==70.2.0
    # via pip-tools
shellingham==1.5.4
    # via typer
shiny==0.10.2
    # via
    #   -r requirements.in
    #   shinywidgets
shinywidgets==0.3.2
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlglot==25.1.0
    # via ibis-framework
stack-data==0.6.3
    # via ipython
starlette==0.37.2
    # via
    #   fastapi
    #   shiny
tenacity==8.4.1
    # via plotly
threadpoolctl==3.5.0
    # via scikit-learn
toolz==0.12.1
    # via ibis-framework
tornado==6.4.1
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   traittypes
traittypes==0.2.1
    # via ipyleaflet
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.1
    # via
    #   fastapi
    #   htmltools
    #   ibis-framework
    #   pydantic
    #   pydantic-core
    #   rsconnect-python
    #   shiny
    #   typer
tzdata==2024.1
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
ujson==5.10.0
    # via fastapi
urllib3==2.2.1
    # via requests
uvicorn==0.30.1
    # via
    #   fastapi
    #   shiny
    #   vetiver
uvloop==0.19.0
    # via uvicorn
vetiver==0.2.4
    # via -r requirements.in
watchfiles==0.22.0
    # via
    #   shiny
    #   uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==12.0
    # via
    #   shiny
    #   uvicorn
wheel==0.43.0
    # via pip-tools
widgetsnbextension==4.0.11
    # via ipywidgets
xxhash==3.4.1
    # via pins
xyzservices==2024.6.0
    # via ipyleaflet
zipp==3.19.2
    # via importlib-metadata
